refactor proxy and interop to there own submodules
